Add option to let user set submit event listener using custom callback
If the form on the page uses JS instead of a standard form submission,
it is necessary for any cc-helper JS to run before default form
submission code.

This is necessary because otherwise modifyOnSubmit (i.e. removing spaces
from cc number) will not run in time and an invalid card number will be
submitted.

To make cc-helper submit event handler run first:

1. Place JS code earlier in the page
1. Ensure submit event is on same element

This code places the submit event on `body`, and the plugin in my use
case places the event on the form itself. In order to stop propagation
of the default submission event on invalid card, I need to place the
cc-helper event on the form itself. In my invocation, I have this new
option set like so:

```javascript
{
  setListenerFormSubmit : function(opts, handler) {
    jQuery(opts.selectors.form).on('submit', handler);
  },
}
```